Off CPU profiling
This is the code that backs
https://github.com/open-telemetry/opentelemetry-ebpf-profiler/pull/144.
It can be reused to add features like requested in
https://github.com/open-telemetry/opentelemetry-ebpf-profiler/issues/33 and
therefore can be an alternative to
https://github.com/open-telemetry/opentelemetry-ebpf-profiler/pull/192.

The idea that enables off CPU profiling is, that perf event and kprobe eBPF
programs are quite similar and can be converted. This allows, with the
dynamic rewrite of tail call maps, the reuse of existing eBPF programs and
concepts.

This proposal adds the new flag '-off-cpu-threshold' that enables off CPU
profiling and attaches the two additional hooks, as discussed in Option B
in https://github.com/open-telemetry/opentelemetry-ebpf-profiler/pull/144.

Outstanding work:
- [ ] Handle off CPU traces in the reporter package
- [ ] Handle off CPU traces in the user space side

Signed-off-by: Florian Lehner <dev@der-flo.net>